requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
cachetools==4.2.4
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.28.0
email-validator==2.3.0
GDAL==3.4.1
gunicorn==23.0.0
httplib2==0.30.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mailersend==2.0.0
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
svgwrite==1.4.3
tomli==2.3.0
Tree==0.2.4
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
zope.interface==8.0.1
